# backend/KokoSkyAPIDOC/kokosky-api-doc-v1.0.0.yml
openapi: 3.0.3
info:
  title: Weather Forecast APIs, KokoSky, Inc.
  version: 1.0.0
  description: |-
    This is the design documentation for KokoSky weather API
  termsOfService: http://kokosky.net/terms
  contact:
    email: contact@kokosky.net
    name: KokoSky, Inc.
    url: https://skyweather.net/contacts
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: http://localhost:{port}
    variables:
      port:
        default: "8080"
    description: This is for development testing
...
paths:
  /api/v1/locations:
    summary: Operations for managing cities and regions around the world (location)
    description: Clients use these APIs to manage locations available for weather forecast
    get:
      summary: Returns a list of managed locations - available for weather forecast
      description: >
        Clients use this API to get a list of managed locations. Each location is uniquely identified by location code, for example 'LACA_US' represents Los Angeles city in California state, in the United States of America (US)
      tags:
        - Location
      operationId: getAllLocations
      responses:
        '200':
          description: Successful operation. There are managed locations available.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
              examples:
                A list of locations:
                  value:
                    - code: LACA_US
                      city_name: Los Angeles
                      region_name: California
                      country_code: US
                      country_name: United States of America
                      enabled: true
                    - code: DELHI_IN
                      city_name: Delhi
                      region_name: Delhi
                      country_code: IN
                      country_name: India
                      enabled: false
                    - code: BWFC_NG
                      city_name: Bwari
                      region_name: Abuja
                      country_code: NG
                      country_name: Nigeria
                      enabled: true
        '204':
          description: No managed locations available
    post:
      summary: Adds a new location to be managed for weather forecast
      description: >
        Clients use this API to put a new location into the system. Location code must be manually specified. Country code and country name based on ISO 3166
      tags:
        - Location
      operationId: addLocation
      requestBody:
        description: Request body for creating a resource
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '201':
          description: >
            Successful added location. Returns details of the newly added location.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: >
            Failed to add location. Request body rejected due to invalid field values
    put:
      summary: Updates an existing location
      description: >
        Clients use this API to modify information of a specific location. Note that location code cannot be changed
      tags:
        - Location
      operationId: updateLocation
      requestBody:
        description: >
          a JSON object that represents the location information that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '200':
          description: >
            Successful update operation. Location updated. Returns details of the recently updated location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: >
            Failed to update location. Request body rejected due to invalid field values
        '404':
          description: No location found with the given code.
  /api/v1/locations{code}:
    summary: Operations for managing a specific location identified by location code
    parameters:
      - $ref: '#/components/parameters/LocationCode'
    get:
      summary: Returns details of a location
      description: Clients use this API to find an existing location in the database by a specific code
      operationId: getLocation
      tags:
        - Location
      responses:
        '200':
          description: >
            Successful retrieved a specific location based on code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          description: No location found with the given code.
    delete:
      summary: Removes an existing location found by given code.
      description: Clients use this API to remove existing location in the database by a specific code
      operationId: deleteLocation
      tags:
        - Location
      responses:
        '204':
          description: Successful removed location.
        '404':
          description: No location found with the given code.
  /api/v1/realtime:
    summary: Operations for accessing realtime weather data based on client's IP address
    get:
      summary: >
        Returns the current weather information of the location based on client's IP address
      description: >
        Clients use this API to get real time weather data of client's location (based on IP address)
      operationId: getRealtimeByIPAddress
      tags:
        - Realtime Weather
      responses:
        '200':
          description: >
            successful retrieval operation. Realtime data available for the client's location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RealtimeWeather'
              examples:
                realtimeWeatherExample:
                  $ref: '#/components/examples/realtimeWeatherExample'
        '400':
          description: Bad request. Could not determine client's IP address
        '404':
          description: No managed location found for the client's IP address
  /api/v1/realtime/{code}:
    summary: >
      Operations for accessing and updating realtime weather data based on location code
    parameters:
      - $ref: '#/components/parameters/LocationCode'
    get:
      summary: Returns the current weather information of a specific location by given code
      description: Clients use this API to get realtime weather data of a specific location by the given code
      operationId: getRealtimeByLocationCode
      tags:
        - Realtime Weather
      responses:
        '200':
          description: Successful retrieval operation. Realtime data available for the given location code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RealtimeWeather'
              examples:
                realtimeWeatherExample:
                  $ref: '#/components/examples/realtimeWeatherExample'
        '404':
          description: No managed location found for the given code
    put:
      summary: Updates realtime weather data based on location code
      description: >-
        Clients use this API to update current weather information for a
        specific location
      operationId: updateRealtimeWeather
      tags:
        - Realtime Weather
      requestBody:
        description: >-
          a JSON object that represents the realtime weather data that needs to be updated. Request body must not include location and last updated time.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RealtimeWeather'
            example:
              temperature: 12
              humidity: 55
              precipitation: 23
              status: Cloudy
              wind_speed: 8
      responses:
        '200':
          description: Successfully updated realtime weather data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RealtimeWeather'
        '404':
          description: No managed location found for the given code
  /api/v1/hourly:
    summary: Operation for accessing hourly weather forecast based on client's IP address
    description: Clients use this API to get hourly forecast information for today
    parameters:
      - $ref: '#/components/parameters/CurrentHour'
    get:
      summary: >-
        Returns weather forecast for upcoming hours in the current day, based on client's IP address
      description: >-
        Clients use this API to get weather forecast for upcoming hours in the current day. Location is determined automatically based on client's IP address.
      operationId: getHourlyForecastByIPAddress
      tags:
        - Hourly Forecast
      responses:
        '200':
          description: >-
            successful retrieval operation. A JSON object representing hourly forecast information available for the client's location - only data for the next hours is returned (based on value of the header X-Current-Hour).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HourlyWeather'
              examples:
                hourlyWeatherExample:
                  $ref: '#/components/examples/hourlyWeatherExample'
        '204':
          description: No hourly forecast data available
        '400':
          description: Error during geolocation process or invalid value of request header X-Current-Hour
        '404':
          description: No managed location found for the client's IP address
  /api/v1/hourly/{code}:
    summary: Operations for accessing and updating hourly weather forecast data based on location code
    description: >-
      Client use this API to get and update hourly weather forecast data based on location code
    parameters:
      - $ref: '#/components/parameters/LocationCode'
    get:
      summary: Returns hourly weather forecast information based on location code.
      description: >-
        Clients use this API to get weather forecast for upcoming hours
        in the current day, for the given location code. Only data for the next hours is returned (based on value of the header X-Current-Hour).
      operationId: getHourlyForecastByLocationCode
      tags:
        - Hourly Forecast
      parameters:
        - $ref: '#/components/parameters/CurrentHour'
      responses:
        '200':
          description: >-
            successful retrieval operation. A JSON object representing hourly forecast data available for the given location code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HourlyWeather'
              examples:
                hourlyWeatherExample:
                  $ref: '#/components/examples/hourlyWeatherExample'
        '204':
          description: No hourly forecast data available
        '400':
          description: Invalid value of request header X-Current-Hour
        '404':
          description: No managed location found for the given location code
    put:
      summary: Updates hourly weather forecast information for a specific location
      description: >-
        Clients use this API to update weather forecast for upcoming hours
        in the current day, for the given location code. Successful update operation will replace all existing data.
      operationId: updateHourlyForecast
      tags:
        - Hourly Forecast
      requestBody:
        description: >-
          a JSON object that represents the hourly weather forecast data that needs to be updated. It must be an array of hourly forecast information as shown in the below example.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HourlyWeather'
            example:
              [
                {
                  "hour_of_day": 8,
                  "temperature": 12,
                  "precipitation": 88,
                  "status": "Cloudy"
                },
                {
                  "hour_of_day": 9,
                  "temperature": 13,
                  "precipitation": 86,
                  "status": "Cloudy"
                },
                {
                  "hour_of_day": 10,
                  "temperature": 15,
                  "precipitation": 60,
                  "status": "Sunny"
                }
              ]
      responses:
        '200':
          description: successful update operation. Hourly weather forecast data updated successfully - all previous data is replaced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HourlyWeather'
              examples:
                hourlyWeatherExample:
                  $ref: '#/components/examples/hourlyWeatherExample'
        '400':
          description: Bad request. Request body contains empty array (no data) or there are some invalid values of fields in hourly forecast information.
        '404':
          description: No managed location found for the given code
...
components:
  schemas:
    Location:
      type: object
      properties:
        code:
          type: string
          description: unique code of a location
          example: LACA_US
        city_name:
          type: string
          description: name of the city
          example: Los Angeles
        region_name:
          type: string
          description: name of the region, state or province
          example: California
        country_code:
          type: string
          description: 2 characters representing country code based on ISO 3166 standard
          example: US
        country_name:
          type: string
          description: name of the country based on ISO 3166 standard
          example: United States of America
        enabled:
          type: boolean
          description: whether the location enabled (true) or disabled (false)
          example: true
    RealtimeWeather:
      type: object
      properties:
        location:
          type: string
          description: >
            name of location, in form of "city name, state/region name, country name"
          example: New York City, New York, United States of America
        temperature:
          type: number
          format: int32
          description: current temperature of the location, default in Celsius degree
          example: 12
        humidity:
          type: number
          format: int32
          description: percentage of humidity
          example: 72
        precipitation:
          type: number
          format: int32
          description: percentage chance of rain
          example: 50
        status:
          type: string
          description: current weather status
          example: Sunny
        wind_speed:
          type: number
          format: int32
          description: current wind speed (default unit is km/h)
          example: 8
        last_updated:
          type: string
          format: date-time
          pattern: yyyy-MM-dd'T'HH:mm:ss'Z'
          description: the last time weather data is updated. Datetime format follows RFC 3339 and ISO 8601.
          example: 2023-02-28T10:50:45Z
    HourlyWeather:
      type: object
      properties:
        location:
          type: string
          description: name of location, in form of "city name, state/region name, country name"
          example: New York, New York, US
        hourly_forecast:
          type: array
          items:
            type: object
            properties:
              hour_of_day:
                type: number
                description: the forecasted hour (in 24h format)
                example: 14
              temperature:
                type: number
                format: int32
                description: forecasted temperature at the hour, default in Celcius degree
                example: 19
              precipitation:
                type: number
                format: int32
                description: percentage chance of rain
                example: 67
              status:
                type: string
                description: current weather status
                example: Cloudy
...
  parameters:
    LocationCode:
      name: code
      in: path
      description: code of location
      required: true
      schema:
        type: string
        example: NYC_US
    CurrentHour:
      name: X-Current-Hour
      in: header
      description: The current hour at client's location
      required: true
      schema:
        type: number
        example: 7
  examples:
    realtimeWeatherExample:
      summary: realtime weather example
      value: >
        {
          "location": "San Francisco, California, United States of America",
          "temperature": 12,
          "humidity": 32,
          "precipitation": 88,
          "status": "Cloudy",
          "wind_speed": 5,
          "last_updated": "2023-02-02T19:23:00Z"
        }
    hourlyWeatherExample:
      summary: hourly weather example
      value: >
        {
          "location": "San Francisco, California, United States of America",
          "hourly_forecast":
          [
            {
              "hour_of_day": 8,            
              "temperature": 12,
              "precipitation": 88,
              "status": "Cloudy"
            },
            {
              "hour_of_day": 9,
              "temperature": 13,
              "precipitation": 86,
              "status": "Cloudy"
            },
            {
              "hour_of_day": 10,
              "temperature": 15,
              "precipitation": 60,
              "status": "Sunny"
            }                
          ]
        
        }
...